chore: move json tracing logic to a new crate
