+++ crates/tracing-json/Cargo.toml
@@ -1,0 +1,19 @@
+lints.workspace = true
+
+[package]
+name = "tracing-json"
+version.workspace = true
+edition.workspace = true
+description.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+documentation.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+serde_json.workspace = true
+tracing.workspace = true
+tracing-subscriber = { workspace = true, features = ["json"] }
+hex.workspace = true
